Add Pipe midleware, add and modify orders endpoints. Task 35.3

+++ src/products/products.controller.ts
@@ -1,6 +1,8 @@
-import { Body, Controller, Delete, Get, Param, Post } from '@nestjs/common';
+import { Body, Controller, Delete, Get, NotFoundException, Param, Post, Put } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDTO } from './dtos/create-product.dto';
+import { ParseUUIDPipe } from '@nestjs/common';
+import { UpdateProductDTO } from './dtos/update-product.dto';
 
 @Controller('products')
 export class ProductsController {
@@ -12,16 +14,30 @@ export class ProductsController {
         return this.productsService.getAll();
     }
     @Get('/:id')
-    getById(@Param('id') id: string) {
+    getById(@Param('id', new ParseUUIDPipe()) id: string) {
+        if (!this.productsService.getById(id))
+            throw new NotFoundException('Product not found');
+
         return this.productsService.getById(id);
     }
     @Delete('/:id')
-    deleteById(@Param('id') id: string) {
+    deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+        if (!this.productsService.getById(id))
+            throw new NotFoundException('Product not found');
+        
         this.productsService.deleteById(id);
         return { success: true };
     }
     @Post('/')
     create(@Body() productData: CreateProductDTO) {
         return this.productsService.create(productData);
+    }
+    @Put('/:id')
+    UpdateById(@Param('id', new ParseUUIDPipe()) id: string, @Body() productData: UpdateProductDTO) {
+        if (!this.productsService.getById(id))
+            throw new NotFoundException('Product not found');
+
+        this.productsService.updateById(id, productData);
+        return { success: true };
     }
 }
